ui(dashboard): polish layout, landing updates, and content modularization
